+++ test/main.rs
@@ -669,6 +669,32 @@ fn main()
         })
     });
 
+    tests.push(GLTest {
+        width: 640,
+        height: 640,
+        name: "image_load_from_raw_pixels_smiley".to_string(),
+        action: Box::new(|renderer| {
+            let image =
+                image::open("test/assets/test_images/smiley_colormap.png").unwrap();
+            let size = image.dimensions();
+
+            renderer.draw_frame(|graphics| {
+                graphics.clear_screen(Color::LIGHT_GRAY);
+
+                let texture = graphics
+                    .create_image_from_raw_pixels(
+                        ImageDataType::RGB,
+                        ImageSmoothingMode::NearestNeighbor,
+                        Vector2::new(size.0, size.1),
+                        &image.to_rgb8()
+                    )
+                    .unwrap();
+
+                graphics.draw_image(Vector2::new(100.0, 100.0), &texture);
+            });
+        })
+    });
+
     for test in tests {
         log::info!("Running test {}", test.name);