+++ CHANGELOG.md
@@ -49,3 +49,8 @@
 * `Graphics2D::draw_image()` is now able to take a tuple as a position
 * When creating an image from raw bytes, the number of bytes is checked
 * Fixed texture load issues where the horizontal byte stride was not a multiple of 4
+
+## 1.1.1
+
+* Now works correctly in Wayland
+* Fixed error when primary monitor is not found